+++ @navikt/core/react/src/overlays/action-menu/ActionMenu.tsx
@@ -1,6 +1,7 @@
 import cl from "clsx";
 import React, { forwardRef, useRef } from "react";
 import { ChevronRightIcon } from "@navikt/aksel-icons";
+import { useModalContext } from "../../modal/Modal.context";
 import { Slot } from "../../slot/Slot";
 import { OverridableComponent, useId } from "../../util";
 import { composeEventHandlers } from "../../util/composeEventHandlers";
@@ -8,7 +9,7 @@ import { createContext } from "../../util/create-context";
 import { useMergeRefs } from "../../util/hooks";
 import { useControllableState } from "../../util/hooks/useControllableState";
 import { requireReactElement } from "../../util/requireReactElement";
-import { Menu } from "../floating-menu/Menu";
+import { Menu, MenuPortalProps } from "../floating-menu/Menu";
 
 /* -------------------------------------------------------------------------- */
 /*                                 ActionMenu                                 */
@@ -20,6 +21,7 @@ type ActionMenuContextValue = {
   open: boolean;
   onOpenChange: (open: boolean) => void;
   onOpenToggle: () => void;
+  rootElement: MenuPortalProps["rootElement"];
 };
 
 const [ActionMenuProvider, useActionMenuContext] =
@@ -29,7 +31,7 @@ const [ActionMenuProvider, useActionMenuContext] =
       "ActionMenu sub-components cannot be rendered outside the ActionMenu component.",
   });
 
-interface ActionMenuProps {
+type ActionMenuProps = {
   children?: React.ReactNode;
   /**
    * Whether the menu is open or not.
@@ -37,14 +39,10 @@ interface ActionMenuProps {
    */
   open?: boolean;
   /**
-   * Whether the menu should be open by default.
-   */
-  defaultOpen?: boolean;
-  /**
    * Callback for when the menu is opened or closed.
    */
   onOpenChange?: (open: boolean) => void;
-}
+} & Pick<MenuPortalProps, "rootElement">;
 
 interface ActionMenuComponent extends React.FC<ActionMenuProps> {
   /**
@@ -247,14 +245,17 @@ interface ActionMenuComponent extends React.FC<ActionMenuProps> {
 const ActionMenuRoot = ({
   children,
   open: openProp,
-  defaultOpen = false,
   onOpenChange,
+  rootElement: rootElementProp,
 }: ActionMenuProps) => {
   const triggerRef = useRef<HTMLButtonElement>(null);
 
+  const modalContext = useModalContext(false);
+  const rootElement = modalContext ? modalContext.ref.current : rootElementProp;
+
   const [open = false, setOpen] = useControllableState({
     value: openProp,
-    defaultValue: defaultOpen,
+    defaultValue: false,
     onChange: onOpenChange,
   });
 
@@ -266,6 +267,7 @@ const ActionMenuRoot = ({
       open={open}
       onOpenChange={setOpen}
       onOpenToggle={() => setOpen((prevOpen) => !prevOpen)}
+      rootElement={rootElement}
     >
       <Menu open={open} onOpenChange={setOpen} modal>
         {children}
@@ -328,58 +330,46 @@ export const ActionMenuTrigger = forwardRef<
 /*                             ActionMenuContent                              */
 /* -------------------------------------------------------------------------- */
 interface ActionMenuContentProps
-  extends Omit<React.HTMLAttributes<HTMLDivElement>, "id">,
-    Pick<React.ComponentPropsWithoutRef<typeof Menu.Portal>, "rootElement"> {
+  extends Omit<React.HTMLAttributes<HTMLDivElement>, "id"> {
   children?: React.ReactNode;
 }
 
 export const ActionMenuContent = forwardRef<
   HTMLDivElement,
   ActionMenuContentProps
->(
-  (
-    {
-      children,
-      className,
-      style,
-      rootElement,
-      ...rest
-    }: ActionMenuContentProps,
-    ref,
-  ) => {
-    const context = useActionMenuContext();
+>(({ children, className, style, ...rest }: ActionMenuContentProps, ref) => {
+  const context = useActionMenuContext();
 
-    return (
-      <Menu.Portal rootElement={rootElement} asChild>
-        <Menu.Content
-          ref={ref}
-          id={context.contentId}
-          aria-labelledby={context.triggerId}
-          className={cl("navds-action-menu__content", className)}
-          {...rest}
-          align="start"
-          sideOffset={4}
-          collisionPadding={10}
-          onCloseAutoFocus={() => {
-            context.triggerRef.current?.focus();
-          }}
-          safeZone={{ anchor: context.triggerRef.current }}
-          style={{
-            ...style,
-            ...{
-              "--__ac-action-menu-content-transform-origin":
-                "var(--ac-floating-transform-origin)",
-              "--__ac-action-menu-content-available-height":
-                "var(--ac-floating-available-height)",
-            },
-          }}
-        >
-          <div className="navds-action-menu__content-inner">{children}</div>
-        </Menu.Content>
-      </Menu.Portal>
-    );
-  },
-);
+  return (
+    <Menu.Portal rootElement={context.rootElement} asChild>
+      <Menu.Content
+        ref={ref}
+        id={context.contentId}
+        aria-labelledby={context.triggerId}
+        className={cl("navds-action-menu__content", className)}
+        {...rest}
+        align="start"
+        sideOffset={4}
+        collisionPadding={10}
+        onCloseAutoFocus={() => {
+          context.triggerRef.current?.focus();
+        }}
+        safeZone={{ anchor: context.triggerRef.current }}
+        style={{
+          ...style,
+          ...{
+            "--__ac-action-menu-content-transform-origin":
+              "var(--ac-floating-transform-origin)",
+            "--__ac-action-menu-content-available-height":
+              "var(--ac-floating-available-height)",
+          },
+        }}
+      >
+        <div className="navds-action-menu__content-inner">{children}</div>
+      </Menu.Content>
+    </Menu.Portal>
+  );
+});
 
 /* -------------------------------------------------------------------------- */
 /*                              ActionMenuLabel                               */
@@ -840,21 +830,17 @@ interface ActionMenuSubProps {
    */
   open?: boolean;
   /**
-   * Whether the sub-menu should be open by default.
-   */
-  defaultOpen?: boolean;
-  /**
    * Callback for when the sub-menu is opened or closed.
    */
   onOpenChange?: (open: boolean) => void;
 }
 
 export const ActionMenuSub = (props: ActionMenuSubProps) => {
-  const { children, open: openProp, onOpenChange, defaultOpen = false } = props;
+  const { children, open: openProp, onOpenChange } = props;
 
   const [open = false, setOpen] = useControllableState({
     value: openProp,
-    defaultValue: defaultOpen,
+    defaultValue: false,
     onChange: onOpenChange,
   });
 
@@ -911,59 +897,48 @@ export const ActionMenuSubTrigger = forwardRef<
 type ActionMenuSubContentElement = React.ElementRef<typeof Menu.Content>;
 
 interface ActionMenuSubContentProps
-  extends React.HTMLAttributes<HTMLDivElement>,
-    Pick<React.ComponentPropsWithoutRef<typeof Menu.Portal>, "rootElement"> {
+  extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
 }
 
 export const ActionMenuSubContent = forwardRef<
   ActionMenuSubContentElement,
   ActionMenuSubContentProps
->(
-  (
-    {
-      children,
-      className,
-      style,
-      rootElement,
-      ...rest
-    }: ActionMenuSubContentProps,
-    ref,
-  ) => {
-    return (
-      <Menu.Portal rootElement={rootElement}>
-        <Menu.SubContent
-          ref={ref}
-          alignOffset={-4}
-          sideOffset={1}
-          collisionPadding={10}
-          {...rest}
-          className={cl(
-            "navds-action-menu__content navds-action-menu__sub-content",
-            className,
-          )}
-          style={{
-            ...style,
-            ...{
-              "--ac-action-menu-content-transform-origin":
-                "var(--ac-floating-transform-origin)",
-              "--ac-action-menu-content-available-width":
-                "var(--ac-floating-available-width)",
-              "--ac-action-menu-content-available-height":
-                "var(--ac-floating-available-height)",
-              "--ac-action-menu-trigger-width":
-                "var(--ac-floating-anchor-width)",
-              "--ac-action-menu-trigger-height":
-                "var(--ac-floating-anchor-height)",
-            },
-          }}
-        >
-          <div className="navds-action-menu__content-inner">{children}</div>
-        </Menu.SubContent>
-      </Menu.Portal>
-    );
-  },
-);
+>(({ children, className, style, ...rest }: ActionMenuSubContentProps, ref) => {
+  const context = useActionMenuContext();
+
+  return (
+    <Menu.Portal rootElement={context.rootElement}>
+      <Menu.SubContent
+        ref={ref}
+        alignOffset={-4}
+        sideOffset={1}
+        collisionPadding={10}
+        {...rest}
+        className={cl(
+          "navds-action-menu__content navds-action-menu__sub-content",
+          className,
+        )}
+        style={{
+          ...style,
+          ...{
+            "--ac-action-menu-content-transform-origin":
+              "var(--ac-floating-transform-origin)",
+            "--ac-action-menu-content-available-width":
+              "var(--ac-floating-available-width)",
+            "--ac-action-menu-content-available-height":
+              "var(--ac-floating-available-height)",
+            "--ac-action-menu-trigger-width": "var(--ac-floating-anchor-width)",
+            "--ac-action-menu-trigger-height":
+              "var(--ac-floating-anchor-height)",
+          },
+        }}
+      >
+        <div className="navds-action-menu__content-inner">{children}</div>
+      </Menu.SubContent>
+    </Menu.Portal>
+  );
+});
 
 /* -------------------------------------------------------------------------- */
 ActionMenu.Trigger = ActionMenuTrigger;