+++ @navikt/core/react/src/util/i18n/locales/en.ts
@@ -1,6 +1,11 @@
 import type { Translations } from "../i18n.types";
 
 export default {
+  global: {
+    showMore: "Show more",
+    showLess: "Show less",
+  },
+
   FileUpload: {
     dropzone: {
       button: "Choose file",
@@ -39,5 +44,23 @@ export default {
   },
   ErrorSummary: {
     Heading: "You must correct the following errors before you can continue:",
+  Loader: {
+    title: "Waiting…",
+  },
+  Modal: {
+    close: "Close",
+  },
+  Pagination: {
+    previous: "Previous",
+    next: "Next",
+  },
+  ProgressBar: {
+    progress: "{current} of {max}",
+    progressUnknown:
+      "Progress is unknown, estimated time is {seconds} seconds.",
+  },
+  Search: {
+    clear: "Clear",
+    search: "Search",
   },
 } satisfies Translations;